refactor: App과 index 분리
